+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Order blogposts by descending publication date within
   (person|category)_detail template
+- Fix filter pane quality issues on handheld devices
 - Fix small quality issues on the course search filters modal
 
 ## [2.4.0] - 2021-04-07